+++ honeybadger/notice.py
@@ -17,14 +17,14 @@ class Notice(object):
         self._process_exception()
 
     def _process_exception(self):
-        if self.exception is None and self.error_class:
+        if self.exception and self.error_message:
+            self.context["error_message"] = self.error_message
+
+        if self.exception is None:
             self.exception = {
                 "error_class": self.error_class,
+                "error_message": self.error_message,
             }
-            if self.error_message:
-                self.exception.update({"error_message": self.error_message})
-        elif self.exception and self.error_message:
-            self.context["error_message"] = self.error_message
 
     @cached_property
     def payload(self):